Fix comment confusion and fix edit states
The player message description said it would send createtrigger message
after creating a camera, but the server code said it would do it in
reverse. Creating a camera first and then defining a zone makes more
sense. It will later be so +attack or something will be used to set the
box points.