Invoke msg thread handlers into main thread
Somehow this never caused any issue but I feel like it definitely will
in the future. HL is all single thread based and the MESSAGE_BEGIN
family functions all operate on the same parts of memory, if our threads
would write at the same time as the engine it would be disaster.